[fix] swizzle NotFound/Content and 301 unmatched /docs/dev/* in .htaccess
The previous fix only swizzled the outer @theme/NotFound, so DocRoot's
fallback (@theme/NotFound/Content) still rendered the default Docusaurus
copy for missing docs under /docs-next/dev/. And legacy /docs/dev/* with
no 1:1 redirect target hit a SPA hydration mismatch and rendered blank.
Now the inner Content is swizzled with the same dev-doc detection, and
unmatched /docs/dev/* (en + zh-CN) is 301'd at the Apache layer before
React boots.
