+++ src/theme/NotFound/index.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { translate } from '@docusaurus/Translate';
+import { PageMetadata } from '@docusaurus/theme-common';
+import Layout from '@theme/Layout';
+import NotFoundContent from '@theme/NotFound/Content';
+
+export default function NotFound() {
+    return (
+        <>
+            <PageMetadata
+                title={translate({
+                    id: 'theme.NotFound.title',
+                    message: 'Page Not Found',
+                })}
+            />
+            <Layout>
+                <NotFoundContent />
+            </Layout>
+        </>
+    );
+}
